+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- internal debug logging of requests to Magento
+
 ### Changed
 - guest checkout now depends on the config 'guest_login_mode'
 
@@ -15,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - pipeline for updating a customer email
 - pipeline for updating a customer
 - support for guest checkout
+
 ### Fixed
 - pipeline for retrieving customer details